Refactor CI: Organize tests folder and xml output
