+++ apps/chat/package.json
@@ -15,7 +15,6 @@
     "ejs": "^2.3.1",
     "express": "^4.12.2",
     "express-session": "^1.10.4",
-    "glob": "^5.0.3",
     "nano": "^6.1.2",
     "node-uuid": "^1.4.3",
     "nodemailer": "^1.3.2",
@@ -33,15 +32,15 @@
     "start": "node server.js",
     "pretest": "npm install",
     "test": "./node_modules/karma/bin/karma start ./tests/unit/frontend/karma.conf.js",
-    "unit-frontend": "./node_modules/karma/bin/karma start ./tests/unit/frontend/karma.conf.js  --single-run",
-    "unit-backend": "node tests/unit/backend/executeJasmineTests.js",
-    "api": "./node_modules/.bin/newman -k -c tests/api_client/collections/testCollection.json.postman_collection --testReportFile tests/api-out.xml",
+    "unit-frontend": "./node_modules/karma/bin/karma start ./tests/unit-frontend/karma.conf.js  --single-run",
+    "unit-backend": "node tests/unit-backend/executeJasmineTests.js",
+    "api": "./node_modules/.bin/newman -k -c tests/api/collections/testCollection.json.postman_collection --testReportFile tests/junit-api.xml",
     "gui": "protractor tests/gui/protractor.conf.js",
     "preupdate-webdriver": "npm install",
     "update-webdriver": "webdriver-manager update",
     "pregui": "npm run update-webdriver",
     "acceptance": "protractor tests/acceptance/protractor.conf.js",
-    "cucumber-junit": "cat ./tests/acceptance/report.json | ./lib/protractor-cucumber-junit/bin/cucumber-junit > tests/acceptance/test-out.xml"
+    "cucumber-junit": "cat ./tests/acceptance/report.json | ./lib/protractor-cucumber-junit/bin/cucumber-junit > tests/junit-acceptance.xml"
   },
   "main": "app.js",
   "repository": {
@@ -63,6 +62,7 @@
     "karma-junit-reporter": "^0.2.2",
     "karma-phantomjs-launcher": "^0.1.4",
     "newman": "^1.2.9",
-    "phantomjs": "^1.9.15"
+    "phantomjs": "^1.9.15",
+    "xml": "~0.0.12"
   }
 }
