+++ apps/chat/tests/acceptance/features/step_definitions/newUserActivationLink_steps.js
@@ -1,6 +1,6 @@
 var assert = require('assert');
-var CAM_MOCKS = require('../../../mock-data.js');
-var protractorHelpers = require('../../../protractorHelpers.js')(browser,element);
+var CAM_MOCKS = require('../../../common/mock-data.js');
+var protractorHelpers = require('../../../common/protractor-helpers.js')(browser,element);
 
 module.exports = function(){
