+++ apps/chat/tests/acceptance/features/support/world.js
@@ -1,5 +1,5 @@
 // features/support/world.js
-var CAM_MOCKS = require('../../../mock-data.js');
+var CAM_MOCKS = require('../../../common/mock-data.js');
 var request = require('request');
 var fs = require('fs');
