+++ apps/chat/tests/api/.jenkins.sh
@@ -5,7 +5,7 @@ export NVM_DIR="/home/ubuntu/.nvm"
 nvm use 0.10
 #nvm alias default 0.11
 
-~/vert.x-2.1.5/bin/vertx run server.vertx.js &
+~/vert.x-2.1.5/bin/vertx run server.vertx.js
 
 sleep 10
 
@@ -15,7 +15,8 @@ cd couchdb
 node populate-couchdb-mocks.js
 cd ..
 
-sudo npm run start -- $1 &
+# Redirect stdout, but not stderr:
+sudo npm run start -- $1 > server.log
 
 sleep 5
