+++ apps/chat/tests/gui/specs/stage.spec.js
@@ -1,4 +1,4 @@
-var changeIdentity = require('../../protractorHelpers.js')(browser,element).changeIdentity;
+var changeIdentity = require('../../common/protractor-helpers.js')(browser,element).changeIdentity;
 
 describe("Main Stage",function  () {
   browser.get('/');
