+++ apps/chat/tests/unit-backend/.jenkins.sh
