+++ apps/chat/tests/unit-backend/spec/support/jasmine.json
@@ -1,5 +1,5 @@
 {
-  "spec_dir": "tests/unit/backend/spec",
+  "spec_dir": "tests/unit-backend/spec",
   "spec_files": [
     "**/*[sS]pec.js"
   ],
