+++ apps/chat/tests/unit-frontend/.jenkins.sh
