+++ apps/chat/tests/unit-frontend/karma.conf.js
@@ -2,7 +2,7 @@ var path = require('path');
 module.exports = function(config){
   config.set({
 
-    basePath : '../../../web/',
+    basePath : '../../web/',
 
     files : [
      'bower_components/angular/angular.js',
@@ -13,8 +13,8 @@ module.exports = function(config){
       'bower_components/jquery/dist/jquery.min.js',
       'js/*.js',
       'js/components/**/*.js',
-      '../tests/unit/frontend/spec/*.js',
-      '../tests/mock-data.js',
+      '../tests/unit-frontend/spec/*.js',
+      '../tests/common/mock-data.js',
     ],
 
     autoWatch : true,
@@ -32,7 +32,7 @@ module.exports = function(config){
     reporters: ['progress','junit'],
     
     junitReporter : {
-      outputFile: path.join(__dirname, '..', '..', 'unit-frontend-out.xml'),
+      outputFile: path.join(__dirname, '..', 'junit-unit-frontend.xml'),
       suite: ''
     },
