+++ apps/chat/tests/unit-frontend/spec/admin_controllers_test.js
